+++ terra-glide/TerraGlide.h
@@ -34,7 +34,7 @@ public:
     TerraGlide& operator=(const TerraGlide&) = default;
 
     TerraGlideStatus setup() noexcept;
-    TerraGlideStatus frame(const Viewport& viewport, double duration) noexcept;
+    TerraGlideStatus frame(const Viewport& viewport, float duration) noexcept;
     void keyDown(int key) noexcept;
     void keyUp(int key) noexcept;
     void leftButtonDown(double xpos, double ypos) noexcept;
